US110605 Switch to use function export instead of mixin
This isn't really doing any UI stuff, so it can just be a standalone function.
